requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
Flask-Admin>=1.6.0
Flask-Login>=0.6.0
Flask-WTF>=1.2.0
Flask-Babel>=3.0.0
Pillow>=10.0.0
python-slugify>=8.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.9
SQLAlchemy[mypy]>=2.0.31
